Fixing URIs
Some of the URIs in the ontology were broken or deprecated. Most importantly the ontology URI itself ended in "/" because the prefix plot: was used for the ont but the ont URI does not equal the namespace URI! Removed "/" and updated org URIs to not use catalogue.linked.data.gov.au, just linked.data.gov.au. Also HTTP -> HTTPS for linked.data.gov.au